+++ src/pages/DirectMessagesPage.js
@@ -1,0 +1,9 @@
+import Header from "../components/header/Header";
+import {Conversations} from "../Conversation";
+
+export function DirectMessagesPage() {
+    return <>
+        <Header hidePostButton={true}/>
+        <Conversations/>
+    </>;
+}